LayoutTests/imported/w3c/web-platform-tests/css/css-grid/subgrid/subgrid-baseline-005-expected.txt








PASS .item 1
PASS .item 2
PASS .item 3
PASS .item 4
PASS .item 5
PASS .item 6
PASS .item 7
FAIL .item 8 assert_equals:
<div data-offset-y="110" class="item small last"></div>
offsetTop expected 110 but got 120
FAIL .item 9 assert_equals:
<div data-offset-y="168" class="item small first"></div>
offsetTop expected 168 but got 158
FAIL .item 10 assert_equals:
<div data-offset-y="243" class="item small last"></div>
offsetTop expected 243 but got 270
FAIL .item 11 assert_equals:
<div data-offset-y="313" class="item small first"></div>
offsetTop expected 313 but got 308
FAIL .item 12 assert_equals:
<div data-offset-y="385" class="item small last"></div>
offsetTop expected 385 but got 420
